Replace mdnsresponder with avahi
