+++ src/freenas/etc/rc.conf
@@ -48,7 +48,7 @@ early_kld_list="dtraceall geom_multipath"
 kld_list="hwpmc t3_tom t4_tom"
 
 dbus_enable="YES"
-mdnsd_enable="YES"
+avahi_daemon_enable="YES"
 
 # Turbo boost
 performance_cpu_freq="HIGH"
