+++ src/middlewared/middlewared/etc_files/local/avahi/avahi-daemon.conf
@@ -1,0 +1,52 @@
+# avahi is free software; you can redistribute it and/or modify it
+# under the terms of the GNU Lesser General Public License as
+# published by the Free Software Foundation; either version 2 of the
+# License, or (at your option) any later version.
+#
+# avahi is distributed in the hope that it will be useful, but WITHOUT
+# ANY WARRANTY; without even the implied warranty of MERCHANTABILITY
+# or FITNESS FOR A PARTICULAR PURPOSE. See the GNU General Public
+# License for more details.
+#
+# You should have received a copy of the GNU Lesser General Public
+# License along with avahi; if not, write to the Free Software
+# Foundation, Inc., 59 Temple Place, Suite 330, Boston, MA 02111-1307
+# USA.
+# See avahi-daemon.conf(5).
+<%
+    import os
+    import sys
+    logger = middleware.logger
+
+    def ipv4_enabled():
+        interfaces = middleware.call_sync('interface.query')
+        for i in interfaces:
+            for alias in i['state']['aliases']:
+                if alias['type'] == 'INET':
+                    return True
+
+        return False
+
+    def ipv6_enabled():
+        interfaces = middleware.call_sync('interface.query')
+        for i in interfaces:
+            for alias in i['state']['aliases']:
+                if alias['type'] == 'INET6':
+                    return True
+
+        return False
+%>
+
+[server]
+use-ipv4=${"yes" if ipv4_enabled() else "no"}
+use-ipv6=${"yes" if ipv6_enabled() else "no"}
+ratelimit-interval-usec=1000000
+ratelimit-burst=1000
+enable-dbus=no
+
+[wide-area]
+enable-wide-area=yes
+
+[publish]
+publish-hinfo=no
+publish-workstation=no
